Move TuyaOAuth2App.js to app.js
